+++ src/Fields/FlexibleContent.php
@@ -15,7 +15,7 @@ namespace Extended\ACF\Fields;
 
 use Extended\ACF\Fields\Settings\ButtonLabel;
 use Extended\ACF\Fields\Settings\ConditionalLogic;
-use Extended\ACF\Fields\Settings\Instructions;
+use Extended\ACF\Fields\Settings\HelperText;
 use Extended\ACF\Fields\Settings\Required;
 use Extended\ACF\Fields\Settings\Wrapper;
 
@@ -23,7 +23,7 @@ class FlexibleContent extends Field
 {
     use ButtonLabel;
     use ConditionalLogic;
-    use Instructions;
+    use HelperText;
     use Required;
     use Wrapper;