+++ tests/Fields/ButtonGroupTest.php
@@ -18,7 +18,7 @@ use Extended\ACF\Tests\Fields\Settings\Choices;
 use Extended\ACF\Tests\Fields\Settings\ConditionalLogic;
 use Extended\ACF\Tests\Fields\Settings\DefaultValue;
 use Extended\ACF\Tests\Fields\Settings\DirectionLayout;
-use Extended\ACF\Tests\Fields\Settings\Instructions;
+use Extended\ACF\Tests\Fields\Settings\HelperText;
 use Extended\ACF\Tests\Fields\Settings\Required;
 use Extended\ACF\Tests\Fields\Settings\Wrapper;
 use InvalidArgumentException;
@@ -29,7 +29,7 @@ class ButtonGroupTest extends FieldTestCase
     use ConditionalLogic;
     use DefaultValue;
     use DirectionLayout;
-    use Instructions;
+    use HelperText;
     use Required;
     use Wrapper;