+++ tests/Fields/LinkTest.php
@@ -15,7 +15,7 @@ namespace Extended\ACF\Tests\Fields;
 
 use Extended\ACF\Fields\Link;
 use Extended\ACF\Tests\Fields\Settings\ConditionalLogic;
-use Extended\ACF\Tests\Fields\Settings\Instructions;
+use Extended\ACF\Tests\Fields\Settings\HelperText;
 use Extended\ACF\Tests\Fields\Settings\Required;
 use Extended\ACF\Tests\Fields\Settings\Wrapper;
 use InvalidArgumentException;
@@ -23,7 +23,7 @@ use InvalidArgumentException;
 class LinkTest extends FieldTestCase
 {
     use ConditionalLogic;
-    use Instructions;
+    use HelperText;
     use Required;
     use Wrapper;