+++ tests/Fields/PageLinkTest.php
@@ -16,7 +16,7 @@ namespace Extended\ACF\Tests\Fields;
 use Extended\ACF\Fields\PageLink;
 use Extended\ACF\Tests\Fields\Settings\ConditionalLogic;
 use Extended\ACF\Tests\Fields\Settings\FilterBy;
-use Extended\ACF\Tests\Fields\Settings\Instructions;
+use Extended\ACF\Tests\Fields\Settings\HelperText;
 use Extended\ACF\Tests\Fields\Settings\Multiple;
 use Extended\ACF\Tests\Fields\Settings\Nullable;
 use Extended\ACF\Tests\Fields\Settings\Required;
@@ -26,7 +26,7 @@ class PageLinkTest extends FieldTestCase
 {
     use ConditionalLogic;
     use FilterBy;
-    use Instructions;
+    use HelperText;
     use Multiple;
     use Nullable;
     use Required;