+++ tests/Fields/PasswordTest.php
@@ -16,8 +16,8 @@ namespace Extended\ACF\Tests\Fields;
 use Extended\ACF\Fields\Password;
 use Extended\ACF\Tests\Fields\Settings\Affixable;
 use Extended\ACF\Tests\Fields\Settings\ConditionalLogic;
+use Extended\ACF\Tests\Fields\Settings\HelperText;
 use Extended\ACF\Tests\Fields\Settings\Immutable;
-use Extended\ACF\Tests\Fields\Settings\Instructions;
 use Extended\ACF\Tests\Fields\Settings\Placeholder;
 use Extended\ACF\Tests\Fields\Settings\Required;
 use Extended\ACF\Tests\Fields\Settings\Wrapper;
@@ -26,8 +26,8 @@ class PasswordTest extends FieldTestCase
 {
     use Affixable;
     use ConditionalLogic;
+    use HelperText;
     use Immutable;
-    use Instructions;
     use Placeholder;
     use Required;
     use Wrapper;