+++ tests/Fields/RepeaterTest.php
@@ -18,7 +18,7 @@ use Extended\ACF\Fields\Text;
 use Extended\ACF\Tests\Fields\Settings\ButtonLabel;
 use Extended\ACF\Tests\Fields\Settings\ConditionalLogic;
 use Extended\ACF\Tests\Fields\Settings\Fields;
-use Extended\ACF\Tests\Fields\Settings\Instructions;
+use Extended\ACF\Tests\Fields\Settings\HelperText;
 use Extended\ACF\Tests\Fields\Settings\Layout;
 use Extended\ACF\Tests\Fields\Settings\Required;
 use Extended\ACF\Tests\Fields\Settings\Wrapper;
@@ -28,7 +28,7 @@ class RepeaterTest extends FieldTestCase
     use ButtonLabel;
     use ConditionalLogic;
     use Fields;
-    use Instructions;
+    use HelperText;
     use Layout;
     use Required;
     use Wrapper;